added utlity
