+++ crates/evm/src/eth/block.rs
@@ -1,7 +1,5 @@
 //! Ethereum block executor.
 
-use alloc::collections::BTreeMap;
-
 use super::{
     dao_fork, eip6110,
     receipt_builder::{AlloyReceiptBuilder, ReceiptBuilder, ReceiptBuilderCtx},
@@ -19,9 +17,8 @@ use crate::{
 };
 use alloc::{borrow::Cow, boxed::Box, vec::Vec};
 use alloy_block_access_list::{
-    balance_change::BalanceChange, code_change::CodeChange, nonce_change::NonceChange,
-    AccountChanges, BlockAccessIndex, BlockAccessList, SlotChanges, StorageChange, MAX_CODE_SIZE,
-    MAX_TXS,
+    balance_change::BalanceChange, AccountChanges, BlockAccessIndex, BlockAccessList, SlotChanges,
+    StorageChange,
 };
 use alloy_consensus::{Header, Transaction, TxReceipt};
 use alloy_eips::{
@@ -32,12 +29,8 @@ use alloy_eips::{
 use alloy_hardforks::EthereumHardfork;
 use alloy_primitives::{Address, Log, B256, U256};
 use revm::{
-    context::result::ExecutionResult,
-    context_interface::result::ResultAndState,
-    database::State,
-    primitives::{StorageKey, StorageValue},
-    state::Account,
-    DatabaseCommit, Inspector,
+    context::result::ExecutionResult, context_interface::result::ResultAndState, database::State,
+    primitives::StorageKey, DatabaseCommit, Inspector,
 };
 
 /// Context for Ethereum block execution.
@@ -287,7 +280,7 @@ where
                     if let Some(created_address) = result.created_address() {
                         if let Some(acc) = state.get(&created_address) {
                             if let Some(bal) = self.block_access_list.as_mut() {
-                                bal.push(from_account_with_tx_index(
+                                bal.push(crate::eth::utils::from_account_with_tx_index(
                                     created_address,
                                     self.receipts.len() as u64,
                                     acc,
@@ -306,7 +299,7 @@ where
                 alloy_primitives::TxKind::Call(address) => {
                     if let Some(acc) = state.get(&address) {
                         if let Some(bal) = self.block_access_list.as_mut() {
-                            bal.push(from_account_with_tx_index(
+                            bal.push(crate::eth::utils::from_account_with_tx_index(
                                 address,
                                 self.receipts.len() as u64,
                                 acc,
@@ -324,7 +317,7 @@ where
             }
             if let Some(acc) = state.get(tx.signer()) {
                 if let Some(bal) = self.block_access_list.as_mut() {
-                    bal.push(from_account_with_tx_index(
+                    bal.push(crate::eth::utils::from_account_with_tx_index(
                         *tx.signer(),
                         self.receipts.len() as u64,
                         acc,
@@ -344,7 +337,7 @@ where
                     let addr = item.address;
                     if state.contains_key(&addr) {
                         if let Some(bal) = self.block_access_list.as_mut() {
-                            bal.push(from_account_with_tx_index(
+                            bal.push(crate::eth::utils::from_account_with_tx_index(
                                 addr,
                                 self.receipts.len() as u64,
                                 state.get(&addr).unwrap(),
@@ -432,18 +425,22 @@ where
                     post_consolidation.push(value);
                 }
 
-                post_system_acc_changes.push(build_post_execution_system_contract_account_change(
-                    WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
-                    pre_withdrawal,
-                    post_withdrawal,
-                    post_system_tx as BlockAccessIndex,
-                ));
-                post_system_acc_changes.push(build_post_execution_system_contract_account_change(
-                    CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS,
-                    pre_consolidation,
-                    post_consolidation,
-                    post_system_tx as BlockAccessIndex,
-                ));
+                post_system_acc_changes.push(
+                    super::utils::build_post_execution_system_contract_account_change(
+                        WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
+                        pre_withdrawal,
+                        post_withdrawal,
+                        post_system_tx as BlockAccessIndex,
+                    ),
+                );
+                post_system_acc_changes.push(
+                    super::utils::build_post_execution_system_contract_account_change(
+                        CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS,
+                        pre_consolidation,
+                        post_consolidation,
+                        post_system_tx as BlockAccessIndex,
+                    ),
+                );
             }
             requests
         } else {
@@ -528,7 +525,7 @@ where
                 receipts: self.receipts,
                 requests,
                 gas_used: self.gas_used,
-                block_access_list: Some(sort_and_remove_duplicates_in_bal(
+                block_access_list: Some(super::utils::sort_and_remove_duplicates_in_bal(
                     self.block_access_list.unwrap(),
                 )),
             },
@@ -613,226 +610,6 @@ where
     {
         EthBlockExecutor::new(evm, ctx, &self.spec, &self.receipt_builder)
     }
-}
-
-/// An utility function for system contract storage allocation.
-pub fn build_post_execution_system_contract_account_change(
-    address: Address,
-    pre: Vec<StorageValue>,
-    post: Vec<StorageValue>,
-    tx_index: BlockAccessIndex,
-) -> AccountChanges {
-    let mut account_changes = AccountChanges::new(address);
-
-    for (i, (pre_val, post_val)) in pre.into_iter().zip(post.into_iter()).enumerate() {
-        let slot = StorageKey::from(i as u64);
-
-        if pre_val != post_val {
-            let change = StorageChange { block_access_index: tx_index, new_value: post_val.into() };
-            account_changes
-                .storage_changes
-                .push(SlotChanges::default().with_slot(slot.into()).with_change(change));
-        } else {
-            account_changes.storage_reads.push(slot.into());
-        }
-    }
-
-    account_changes
-}
-
-/// An utility function to build block access list with tx_index
-pub fn from_account_with_tx_index(
-    address: Address,
-    block_access_index: u64,
-    account: &Account,
-) -> AccountChanges {
-    let mut account_changes = AccountChanges::default();
-
-    for key in &account.storage_access.reads {
-        tracing::debug!("Storage read at {:#x}: {:#x} ", address, key);
-        account_changes.storage_reads.push((*key).into());
-    }
-
-    // Group writes by slots
-    let mut slot_map: BTreeMap<StorageKey, Vec<StorageChange>> = BTreeMap::new();
-
-    for (slot, (_pre, post)) in &account.storage_access.writes {
-        tracing::debug!("Storage write at {:#x}: {:#x} -> {:#x}", address, slot, post);
-        slot_map
-            .entry(*slot)
-            .or_default()
-            .push(StorageChange { block_access_index, new_value: (*post).into() });
-    }
-
-    // Convert slot_map into SlotChanges and push into account_changes
-    for (slot, changes) in slot_map {
-        account_changes.storage_changes.push(SlotChanges { slot: slot.into(), changes });
-    }
-
-    // False if zero value transfer
-    if !account.balance_change.1 {
-        account_changes
-            .balance_changes
-            .push(BalanceChange { block_access_index, post_balance: account.balance_change.0 });
-    }
-
-    let (pre_nonce, post_nonce) = account.nonce_change;
-    if pre_nonce != post_nonce {
-        account_changes
-            .nonce_changes
-            .push(NonceChange { block_access_index, new_nonce: post_nonce });
-    }
-
-    let code = &account.code_change;
-    if !code.is_empty() {
-        account_changes
-            .code_changes
-            .push(CodeChange { block_access_index, new_code: code.clone() });
-    }
-
-    account_changes.address = address;
-    account_changes
-}
-
-/// Sort block-level access list and removes duplicates entries by merging them together.
-pub fn sort_and_remove_duplicates_in_bal(mut bal: BlockAccessList) -> BlockAccessList {
-    tracing::debug!("Bal before sort: {:#?}", bal);
-    bal.sort_by_key(|ac| ac.address);
-    let mut merged: Vec<AccountChanges> = Vec::new();
-
-    for account in bal {
-        if let Some(last) = merged.last_mut() {
-            if last.address == account.address {
-                // Same address → extend fields
-                last.storage_changes.extend(account.storage_changes);
-                last.storage_reads.extend(account.storage_reads);
-                last.balance_changes.extend(account.balance_changes);
-                last.nonce_changes.extend(account.nonce_changes);
-                last.code_changes.extend(account.code_changes);
-                continue;
-            }
-        }
-        merged.push(account);
-    }
-    tracing::debug!("Bal after sort: {:#?}", merged);
-    merged
-}
-
-/// Validates a Block Access List against execution constraints.
-pub fn validate_block_access_list_against_execution(block_access_list: &BlockAccessList) -> bool {
-    // 1. Validate structural constraints
-    for account in block_access_list {
-        let changed_slots: alloy_primitives::map::HashSet<_> =
-            account.storage_changes.iter().map(|sc| B256::from(sc.slot)).collect();
-        let read_slots: alloy_primitives::map::HashSet<_> =
-            account.storage_reads.iter().cloned().collect();
-
-        // A slot should not be in both changes and reads (per EIP-7928)
-        if !changed_slots.is_disjoint(&read_slots) {
-            return false;
-        }
-    }
-
-    // 2. Validate ordering (addresses should be sorted lexicographically)
-    let addresses: Vec<_> = block_access_list.iter().map(|account| account.address).collect();
-    let mut sorted_addresses = addresses.clone();
-    sorted_addresses.sort();
-    if addresses != sorted_addresses {
-        return false;
-    }
-
-    // 3. Validate all data is within bounds
-    let max_block_access_index = MAX_TXS + 1; // 0 for pre-exec, 1..MAX_TXS for txs, MAX_TXS+1 for post-exec
-    for account in block_access_list {
-        // Validate storage slots are sorted within each account
-        let storage_slots: Vec<_> = account.storage_changes.iter().map(|sc| sc.slot).collect();
-        let mut sorted_storage_slots = storage_slots.clone();
-        sorted_storage_slots.sort();
-        if storage_slots != sorted_storage_slots {
-            return false;
-        }
-
-        // Check storage changes
-        for slot_changes in &account.storage_changes {
-            // Check changes are sorted by block_access_index
-            let indices: Vec<_> =
-                slot_changes.changes.iter().map(|c| c.block_access_index).collect();
-            let mut sorted_indices = indices.clone();
-            sorted_indices.sort();
-            if indices != sorted_indices {
-                return false;
-            }
-
-            for change in &slot_changes.changes {
-                if change.block_access_index > max_block_access_index.try_into().unwrap() {
-                    return false;
-                }
-            }
-        }
-
-        // Check balance changes are sorted by block_access_index
-        let balance_indices: Vec<_> =
-            account.balance_changes.iter().map(|bc| bc.block_access_index).collect();
-        let mut sorted_balance_indices = balance_indices.clone();
-        sorted_balance_indices.sort();
-        if balance_indices != sorted_balance_indices {
-            return false;
-        }
-
-        for balance_change in &account.balance_changes {
-            if balance_change.block_access_index > max_block_access_index.try_into().unwrap() {
-                return false;
-            }
-        }
-
-        // Check nonce changes are sorted by block_access_index
-        let nonce_indices: Vec<_> =
-            account.nonce_changes.iter().map(|nc| nc.block_access_index).collect();
-        let mut sorted_nonce_indices = nonce_indices.clone();
-        sorted_nonce_indices.sort();
-        if nonce_indices != sorted_nonce_indices {
-            return false;
-        }
-
-        for nonce_change in &account.nonce_changes {
-            if nonce_change.block_access_index > max_block_access_index.try_into().unwrap() {
-                return false;
-            }
-        }
-
-        // Check code changes are sorted by block_access_index
-        let code_indices: Vec<_> =
-            account.code_changes.iter().map(|cc| cc.block_access_index).collect();
-        let mut sorted_code_indices = code_indices.clone();
-        sorted_code_indices.sort();
-        if code_indices != sorted_code_indices {
-            return false;
-        }
-
-        for code_change in &account.code_changes {
-            if code_change.block_access_index > max_block_access_index.try_into().unwrap() {
-                return false;
-            }
-            if code_change.new_code.len() > MAX_CODE_SIZE {
-                return false;
-            }
-        }
-    }
-
-    true
-}
-
-/// Validate the block access list against an expected block access list
-pub fn validate_block_access_list(
-    block_access_list: &BlockAccessList,
-    expected_block_access_list: &BlockAccessList,
-) -> bool {
-    if alloy_primitives::keccak256(alloy_rlp::encode(block_access_list))
-        != alloy_primitives::keccak256(alloy_rlp::encode(expected_block_access_list))
-    {
-        return false;
-    }
-    true
 }
 
 #[cfg(test)]
